feat: memoize cache option
