+++ packages/utils/src/lodash.test.ts
@@ -10,6 +10,7 @@ import {
   mapGroupBy,
   mapKeys,
   mapValues,
+  memoize,
   objectKeys,
   objectOmit,
   objectOmitBy,
@@ -377,6 +378,75 @@ describe("once", () => {
     expect(count).toMatchInlineSnapshot("2");
     expect(f(4)).toMatchInlineSnapshot("2");
     expect(count).toMatchInlineSnapshot("2");
+  });
+});
+
+describe(memoize, () => {
+  it("basic", () => {
+    const f = vi.fn().mockImplementation((v: number) => v + 10);
+    const g = memoize(f);
+    expect(g(1)).toMatchInlineSnapshot("11");
+    expect(g(1)).toMatchInlineSnapshot("11");
+    expect(f.mock.calls).toMatchInlineSnapshot(`
+      [
+        [
+          1,
+        ],
+      ]
+    `);
+  });
+
+  it("'undefined' not cached", () => {
+    const f = vi
+      .fn()
+      .mockImplementation((v: number) => (v > 0 ? v + 10 : undefined));
+    const g = memoize(f);
+    expect(g(0)).toMatchInlineSnapshot("undefined");
+    expect(g(0)).toMatchInlineSnapshot("undefined");
+    expect(g(1)).toMatchInlineSnapshot("11");
+    expect(g(1)).toMatchInlineSnapshot("11");
+    expect(f.mock.calls).toMatchInlineSnapshot(`
+      [
+        [
+          0,
+        ],
+        [
+          0,
+        ],
+        [
+          1,
+        ],
+      ]
+    `);
+  });
+
+  it("object", () => {
+    // from https://lodash.com/docs/4.17.15#memoize
+
+    const object = { a: 1, b: 2 };
+    const other = { c: 3, d: 4 };
+
+    const f = memoize(Object.values);
+    expect(f(object)).toMatchInlineSnapshot(`
+      [
+        1,
+        2,
+      ]
+    `);
+    expect(f(other)).toMatchInlineSnapshot(`
+      [
+        3,
+        4,
+      ]
+    `);
+
+    object.a = 2;
+    expect(f(object)).toMatchInlineSnapshot(`
+      [
+        1,
+        2,
+      ]
+    `);
   });
 });
